Allow for specifing node names for bolt
Prior to this commit, the nodes had to be specified the node list in URI
format. This commit adds the ability to specify node names in an array
to be added by their connection info to the node list URI. It allows for
specifying `all` for all nodes, an exclude list, and an include list.