Default to a new bolt config object
Prior to this commit, the default bolt config object came from the
machine's bolt config. With the recent change to split the config object
into a provisioner and a global, the global would be used for the VM
object. As this did not have all of the required options, triggers would
fail. This commit defaults to creating a new bolt config object in the
runner.