Fix an override issue with config_builder
Prior to this commit, there were two classes that monkey patched
the same config_builder classes. This resulted in only one of the two
working and the other silently failing. This commit moves the monkey
patches into a seperate file and creates a module to be included in both
of the origional classes. This commit fixes the issue where VM level
config items did not make it into the inventory.