Move methods to util
Prior to this commit there were some methods in runner that should have
been in the util. This commit moves the method for running a command
locally on the platform to util. This commit moves the definition of the
inventory_path to a method in the util module. These changes lay the
ground work for adding commands into vagrant bolt.