Catch generic ready exceptions
Prior to this commit, we only rescued on a specific execption when
checking if a machine is ready. This commit updates it to catach all
exceptions and return that the node is not ready. This will prevent
issues when there is a DNS issue with specific nodes in the vmpooler
provider.