(SUP-1231) Remove modulepath default
Prior to this commit, the modulepath defaulted to `modules`. This caused
the `modulepath` to be appended to each of the bolt commands. This is
not necessary as the bolt behavior will look for modules in the
`site-modules` directory in the `boltdir` by default. This commit
removes the default and disables the `modulepath` from the command
output as well as tasks and plans.