(SUP-1231) Convert absolute to relative paths
Prior to this commit, we used the absolute path in bolt command output.
This was done because vagrant commands can be run in any subdirectory
inside the vagrant directory, and if the paths were sent in as they were
configured, you would only be able to use the bolt plugin at the root
of the vagrant tree. The absolute paths became too long and could
eventually hit a limit on the number of characters for a command, plus
they were not great to look at.

This commit changes the behavior to calcuate the paths to the files
based on the vagrant root path and the current working directory and
then inserts the relative path into the command. For example:

Root dir: `/root/path'
CWD:      `/root/path/subdir/subdir2`

With the paths above, the path to the `boltdir` by default would be
`../../`.