Default to using the full path for bolt
In some cases, where bolt has been installed as a gem the PATH will not
have `/opt/puppetlabs/bin` first, which allows for running the gem
installed verion of bolt. This version does not unset the GEMPATH and
other ENV variables, which causes failures when running under vagrant.

This commit adds some basic logic to try to use
`/opt/puppetlabs/bin/bolt` if it exists.