Add a machine_alias parameter
Prior to this commit, we automatically set the alias in the bolt config.
With the addition of the `machine_name` it was possible to get a name
conflict if both used the machine name, which caused errors in running
bolt. This commit introduces the new parameter and configures the
`machine_name` to default to the machine name. If the `machine_name` is
configured to a different name, the alias will default to the machine
name. This logic enables the bolt plugin to be able to reference the
machine by name, while enabling target naming.