(maint) Remove run_as from the command line
Prior to this commit, the run_as was always sent on the command line to
bolt. In bolt 2.0 the run_as is overridden by the inventory.yaml. Since
we always have it in the inventory.yaml, the command line argument is
not used. The run-as has been removed in this commit.